moved friendlyDates into Util folder
